Updated the link to Cadence documentation pages in Examples.tsx
Updated the link to Cadence documentation pages in Examples.tsx

Cadence-lang.org seems to have changed the URL for these documentation pages.

For example:
https://cadence-lang.org/docs/tutorial/01-first-steps

Has been changed to:
https://cadence-lang.org/docs/tutorial/first-steps

Therefore, links to the Cadence docs present in the Example.tsx needed to be updated. 

Have removed the numerical part from all the links and confirmed they are working.